rough move functions to files
